requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
frozenlist==1.5.0
idna==3.10
kwork==0.0.5
multidict==6.1.0
propcache==0.2.0
pydantic==1.10.19
python-dotenv==1.0.1
typing_extensions==4.12.2
websockets==9.1
yarl==1.17.1

